Compile fixes

+++ include/threading/transports/core/TQueueTransport.hpp
@@ -86,7 +86,7 @@ public:
     [[nodiscard]] bool posting_poisoned() const noexcept;
     [[nodiscard]] bool post(const T& src) noexcept { return post(&src, 1u); }
     [[nodiscard]] bool post(const T* const src, const std::uint32_t count = 1u) noexcept;
-    [[nodiscard]] bool post(const TPodConstView<T>& src) noexcept { return post(src.data(), src.size()); }
+    [[nodiscard]] bool post(const TPodConstView<T>& src) noexcept { return post(src.data(), static_cast<std::uint32_t>(src.size())); }
     [[nodiscard]] bool post_would_reallocate(const std::uint32_t count) const noexcept;
 
     //  Consumer status and operations
@@ -95,7 +95,7 @@ public:
     [[nodiscard]] bool reading_is_ready() const noexcept;
     [[nodiscard]] bool read(T& dst) noexcept { return read(&dst, 1u); }
     [[nodiscard]] bool read(T* const dst, const std::uint32_t count = 1u) noexcept;
-    [[nodiscard]] bool read(const TPodView<T>& dst) noexcept { return read(dst.data(), dst.size()); }
+    [[nodiscard]] bool read(const TPodView<T>& dst) noexcept { return read(dst.data(), static_cast<std::uint32_t>(dst.size())); }
     [[nodiscard]] std::uint32_t current_readable_count() const noexcept;
     [[nodiscard]] std::uint32_t refresh_readable_count() noexcept;
